requirements.txt
click==7.1.2
cmake==3.18.4.post1
dlib==19.21.1
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==1.1.2
Flask-Cors==3.0.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.4
opencv-python==4.4.0.46
Pillow==8.0.1
six==1.15.0
Werkzeug==1.0.1
